Big files reorganization
